refactor: types into modules
