+++ packages/base/lib/denormalizers.d.ts
@@ -1,0 +1,6 @@
+import core from "./core";
+import { OutPoint, Script } from "./api";
+
+export function DenormalizeOutPoint(outPoint: core.OutPoint): OutPoint;
+
+export function DenormalizeScript(script: core.Script): Script;
