+++ packages/base/lib/helpers.d.ts
@@ -1,0 +1,7 @@
+import { Cell } from "./api";
+import { QueryOptions } from "./indexer";
+
+export declare function isCellMatchQueryOptions(
+  cell: Cell,
+  queryOptions: QueryOptions
+): boolean;
