+++ packages/base/lib/logger.d.ts
@@ -1,0 +1,4 @@
+export type LogLevel = "warn" | "error" | string;
+// Indexer
+export type Logger = (level: LogLevel, message: string) => void;
+export type Log = (message: string) => void;
